+++ src/rtlib/data_readulong.c
@@ -8,11 +8,11 @@ FBCALL void fb_DataReadULongint( unsigned long long *dst )
 
 	if( __fb_data_ptr ) {
 		if( __fb_data_ptr->len == FB_DATATYPE_OFS ) {
-			*dst = (unsigned long long)(unsigned long)__fb_data_ptr->ofs;
+			*dst = (size_t)__fb_data_ptr->ofs;
 		} else if( __fb_data_ptr->len & FB_DATATYPE_WSTR ) {
-			*dst = (unsigned long long)fb_WstrToULongint( __fb_data_ptr->wstr, __fb_data_ptr->len & 0x7FFF );
+			*dst = fb_WstrToULongint( __fb_data_ptr->wstr, __fb_data_ptr->len & 0x7FFF );
 		} else {
-			*dst = (unsigned long long)fb_hStr2ULongint( __fb_data_ptr->zstr, __fb_data_ptr->len );
+			*dst = fb_hStr2ULongint( __fb_data_ptr->zstr, __fb_data_ptr->len );
 		}
 	} else {
 		/* no more DATA */
